Release cross by default.
The fact that you have to type "release cross" for projects that are
configured for cross building, but that "release cross" is the same as
"release" for projects that are not, caused us some problems. In
particular, people would forget to run "release cross".

So with this change, "release" is the same as "release cross". If you
wish to disable this feature for a particular project, you can include a
setting:

  sbtrelease.ReleasePlugin.ReleaseKeys.autoCrossBuild := false

...after pulling in sbtrelease.ReleasePlugin.releaseSettings.