Update tests to new sbt versions
Newer versions of sbt appear to be writing to the file
`project/build.properties` when the sbt shell is started.

This created an untracked file which would block the release
from proceeding.

The fix is to first start sbt and then commit all the files
to the repository afterwards. After the commit, the actual
release test command is executed.